[XLA:GPU] Reuse kernels for fusions to speed up compilation
Main idea: If we have multiple fusion operations with the same computation, it is enough to compile one kernel, a separate kernel is not needed for every such operation.

We reuse kernels for fusions if:
- The computation of the fusion is the same
- And the alignment, aliasing and "constness" of the operands are the same

The calling convention changed this way for fusions:
- We pass all arguments / output buffers separately, not as a single temp_buff. Constant arguments are also passed as arguments to the kernel.
- This means that kernels don't have hardcoded offsets to their argument buffers, they simply receive direct pointers to them.

This change can speed up the part of the compilation after the HLO passes multiple times. For example we have measured 2-24x speedup for different big HLOs. For small HLOs, the difference is not significant.

PiperOrigin-RevId: 512921717